requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.0.1
coreapi==2.3.3
djangorestframework-simplejwt==5.3.0